requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.26.0 